Updating owned algorithms on login + migration from ints to longs

+++ app/src/main/java/fr/zabricraft/delta/activities/EditorActivity.java
@@ -61,7 +61,7 @@ public class EditorActivity extends AppCompatActivity {
         lastInstance = null;
     }
 
-    public void updateRemoteId(Integer remote_id) {
+    public void updateRemoteId(Long remote_id) {
         fragment.updateRemoteId(remote_id);
     }
 }
